Registration Page operational
