--- client/src/components/SignupForm/SignupForm.css
@@ -1,29 +1,0 @@
-.form h3 {
-  color: black;
-  font-size: 2em;
-  text-align: center;
-  padding: 10px 0px 10px 0px;
-}
-
-input.form {
-  display: inline-block;
-}
-
-button.signup {
-  padding: 5px;
-}
-
-h2.homeform {
-  font-size: 24px;
-  font-style: italic;
-  margin-top: 30px;
-  margin-bottom: 10px;
-  color: black;
-  padding-left: 10px;
-  font-weight: bolder;
-  font-family: "Poppins", sans-serif;
-}
-
-.screen {
-  background-color: rgba(255,255,255,0.1);
-  }
