--- client/src/components/SignupForm/index.js
@@ -1,1 +1,0 @@
-export { default } from "./SignupForm";
